# agri_uav_disease_system/backend/docs/integration_package/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "三下乡无人机水稻病虫害识别系统",
    "version": "mvp-1"
  },
  "paths": {
...
    "/api/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts",
        "operationId": "list_alerts_api_alerts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "plot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plot Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/records/{record_id}": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "Get Record",
        "operationId": "get_record_api_records__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mobile/overview": {
      "get": {
        "tags": [
          "mobile"
        ],
        "summary": "Mobile Overview",
        "operationId": "mobile_overview_api_mobile_overview_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mobile/plots": {
      "get": {
        "tags": [
          "mobile"
        ],
        "summary": "Mobile Plots",
        "operationId": "mobile_plots_api_mobile_plots_get",
        "parameters": [
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "region_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Name"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobilePlotListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mobile/plots/{plot_id}": {
      "get": {
        "tags": [
          "mobile"
        ],
        "summary": "Mobile Plot Detail",
        "operationId": "mobile_plot_detail_api_mobile_plots__plot_id__get",
        "parameters": [
          {
            "name": "plot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobilePlotDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mobile/records/{record_id}": {
      "get": {
        "tags": [
          "mobile"
        ],
        "summary": "Mobile Record Detail",
        "operationId": "mobile_record_detail_api_mobile_records__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileRecordDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mobile/suggestions/{record_id}": {
      "get": {
        "tags": [
          "mobile"
        ],
        "summary": "Mobile Suggestion",
        "operationId": "mobile_suggestion_api_mobile_suggestions__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Suggestion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/prediction/plots/{plot_id}": {
      "get": {
        "tags": [
          "prediction"
        ],
        "summary": "Predict Plot",
        "operationId": "predict_plot_api_prediction_plots__plot_id__get",
        "parameters": [
          {
            "name": "plot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plot Id"
            }
          },
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "title": "Window Days"
            }
          },
          {
            "name": "disease",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disease"
            }
          },
          {
            "name": "save",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Save"
            }
          },
          {
            "name": "create_alert",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Create Alert"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskPredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mobile/predictions": {
      "get": {
        "tags": [
          "prediction"
        ],
        "summary": "Mobile Predictions",
        "operationId": "mobile_predictions_api_mobile_predictions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobilePredictionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mobile/plots/{plot_id}/prediction": {
      "get": {
        "tags": [
          "prediction"
        ],
        "summary": "Mobile Plot Prediction",
        "operationId": "mobile_plot_prediction_api_mobile_plots__plot_id__prediction_get",
        "parameters": [
          {
            "name": "plot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskPredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/knowledge/diseases/{disease_id}": {
      "get": {
        "tags": [
          "experimental-knowledge"
        ],
        "summary": "Get Disease",
        "operationId": "get_disease_api_knowledge_diseases__disease_id__get",
        "parameters": [
          {
            "name": "disease_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Disease Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiseaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/search": {
      "post": {
        "tags": [
          "experimental-knowledge"
        ],
        "summary": "Search Knowledge",
        "operationId": "search_knowledge_api_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertDetail"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "AlertPageResponse"
      },
...
      "DetectionResult": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "detection_result"
          },
          "record_id": {
            "type": "string",
            "title": "Record Id"
          },
          "image_id": {
            "type": "string",
            "title": "Image Id"
          },
          "field_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Id"
          },
          "plot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Id"
          },
          "plot_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Name"
          },
          "region_name": {
            "type": "string",
            "title": "Region Name"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "result_image_url": {
            "type": "string",
            "title": "Result Image Url"
          },
          "image_width": {
            "type": "integer",
            "title": "Image Width"
          },
          "image_height": {
            "type": "integer",
            "title": "Image Height"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "detector_mode": {
            "type": "string",
            "title": "Detector Mode"
          },
          "is_smoke": {
            "type": "boolean",
            "title": "Is Smoke",
            "default": false
          },
          "model_stage": {
            "type": "string",
            "title": "Model Stage",
            "default": "mock"
          },
          "formal_metric_available": {
            "type": "boolean",
            "title": "Formal Metric Available",
            "default": false
          },
          "current_target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Target Type"
          },
          "category_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Type"
          },
          "fallback_to_mock": {
            "type": "boolean",
            "title": "Fallback To Mock",
            "default": false
          },
          "model_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Hint"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          },
          "model_stage_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Stage Hint"
          },
          "uav_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uav Task Id"
          },
          "abnormal_region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abnormal Region Id"
          },
          "model_display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Display Name"
          },
          "model_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Warning"
          },
          "model_usage_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Usage Scope"
          },
          "model_capability_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Capability Level"
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type"
          },
          "result_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Type"
          },
          "disease_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Name"
          },
          "model_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Sha256"
          },
          "input_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Config"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "min_area": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Area"
          },
          "disease_area_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease Area Ratio"
          },
          "mask_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mask Url"
          },
          "overlay_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlay Url"
          },
          "probability_map_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability Map Url"
          },
          "production_scope": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Production Scope"
          },
          "human_review_required": {
            "type": "boolean",
            "title": "Human Review Required",
            "default": false
          },
          "human_review_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Review Status"
          },
          "human_review_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Review Label"
          },
          "issue_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issue Tags"
          },
          "reviewer_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Note"
          },
          "alerting_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alerting Enabled"
          },
          "latest_alerts_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Alerts Enabled"
          },
          "active_model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Model Version"
          },
          "geo": {
            "$ref": "#/components/schemas/Geo"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "type": "array",
            "title": "Detections"
          },
          "summary": {
            "$ref": "#/components/schemas/DetectionSummary"
          },
          "suggestion": {
            "$ref": "#/components/schemas/Suggestion"
          }
        },
        "type": "object",
        "required": [
          "record_id",
          "image_id",
          "region_name",
          "timestamp",
          "image_url",
          "result_image_url",
          "image_width",
          "image_height",
          "source_type",
          "model_name",
          "model_version",
          "detector_mode",
          "geo",
          "detections",
          "summary",
          "suggestion"
        ],
        "title": "DetectionResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MobilePlotListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MobilePlotItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "MobilePlotListResponse"
      },
...
    }
  }
}